requirements.txt
# Web Framework
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# Visualization
rerun-sdk>=0.15.0
playwright==1.58.0

# Optional: robotics (if needed for advanced features)
# pinocchio==2.6.20

# Development (optional)
# pytest==7.4.0
# black==23.9.1
